Externalize store functions
